rename e2e scenario
